Show episodes honor ticket access

+++ web-server/database/operation/show_episode.py
@@ -8,14 +8,25 @@ from settings import config
 import database.operation.show as db_show
 import database.operation.show_season as db_season
 
-def get_show_episode_list_by_shelf(shelf_id:int):
+def get_show_episode_list_by_shelf(ticket:dm.Ticket,shelf_id:int):
+    if not ticket.is_allowed(shelf_id=shelf_id):
+        return []
     with DbSession() as db:
-        return (
+        query = (
             db.query(dm.ShowEpisode)
+            .filter(dm.ShowEpisode.season.show.shelf.id == shelf_id)
             .options(sorm.joinedload(dm.ShowEpisode.season))
             .options(sorm.joinedload(dm.ShowEpisode.season.show))
-            .filter(dm.ShowEpisode.season.show.shelf.id == shelf_id).all()
+            .options(sorm.joinedload(dm.ShowEpisode.tags))
         )
+        episodes = query.all()
+        results = []
+        for episode in episodes:
+            if not ticket.is_allowed(tag_provider=episode.get_tag_ids):
+                continue
+            episode.convert_local_paths_to_web_paths(config=config)
+            results.append(episode)
+        return results
 
 def create_show_episode(show_season_id: int, episode_order_counter: int):
     with DbSession() as db:
@@ -28,28 +39,47 @@ def create_show_episode(show_season_id: int, episode_order_counter: int):
         return dbm
 
 
-def get_show_episode_list(show_id:int):
-    seasons = db_show.get_show_season_list_by_show_id(show_id=show_id)
+def get_show_episode_list_by_show(ticket:dm.Ticket,show_id:int):
+    show = db_show.get_show_by_id(ticket=ticket,show_id=show_id)
+    if not show:
+        return []
+    seasons = db_season.get_show_season_list_by_show_id(ticket=ticket,show_id=show_id)
+    if not seasons:
+        return []
     season_ids = [xx.id for xx in seasons]
     with DbSession() as db:
-        return (
+        query = (
             db.query(dm.ShowEpisode)
             .filter(dm.ShowEpisode.show_season_id.in_(season_ids))
-            .all()
+            .options(sorm.joinedload(dm.ShowEpisode.season))
+            .options(sorm.joinedload(dm.ShowEpisode.season.show))
+            .options(sorm.joinedload(dm.ShowEpisode.tags))
         )
-def get_show_episode_by_id(episode_id: int):
+        episodes = query.all()
+        results = []
+        for episode in episodes:
+            if not ticket.is_allowed(tag_provider=episode.get_tag_ids):
+                continue
+            episode.convert_local_paths_to_web_paths(config=config)
+            results.append(episode)
+        return results
+
+def get_show_episode_by_id(ticket:dm.Ticket,episode_id: int):
     with DbSession() as db:
-        episode = (
+        query = (
             db.query(dm.ShowEpisode)
-            .options(sorm.joinedload(dm.ShowEpisode.video_files))
-            .options(
-                sorm.joinedload(dm.ShowEpisode.season)
-                .joinedload(dm.ShowSeason.show)
-                .joinedload(dm.Show.shelf)
-            )
             .filter(dm.ShowEpisode.id == episode_id)
-            .first()
+            .options(sorm.joinedload(dm.ShowEpisode.video_files))
+            .options(sorm.joinedload(dm.ShowEpisode.image_files))
+            .options(sorm.joinedload(dm.ShowEpisode.metadata_files))
+            .options(sorm.joinedload(dm.ShowEpisode.tags))      
         )
+        episode = query.first()
+        season = db_season.get_show_season_by_id(ticket=ticket,season_id=episode.season.id)
+        if not season:
+            return None
+        if not ticket.is_allowed(tag_provider=episode.get_tag_ids):
+            return None
         episode.convert_local_paths_to_web_paths(config=config)
         return episode
 
@@ -64,18 +94,24 @@ def get_show_episode_by_season_order(show_season_id: int, episode_order_counter:
             .first()
         )
 
-
-def get_show_episode_list_by_season(show_season_id: int):
+def get_show_episode_list_by_season(ticket:dm.Ticket,show_season_id: int):
+    season = db_season.get_show_season_by_id(ticket=ticket,season_id=show_season_id)
+    if not season:
+        return []
     with DbSession() as db:
-        query = db.query(dm.ShowEpisode).join(dm.ShowSeason).join(dm.Show)
-        episodes = (
-            query.filter(dm.ShowEpisode.show_season_id == show_season_id)
+        query = (
+            db.query(dm.ShowEpisode)
+            .filter(dm.ShowEpisode.show_season_id == show_season_id)            
             .order_by(dm.ShowEpisode.episode_order_counter)
-            .all()
         )
+        episodes = query.all()
+        results = []
         for episode in episodes:
+            if not ticket.is_allowed(tag_provider=episode.get_tag_ids):
+                continue
             episode.convert_local_paths_to_web_paths(config=config)
-        return episodes
+            results.append(episode)
+        return results
 
 def create_show_episode_video_file(show_episode_id: int, video_file_id: int):
     with DbSession() as db:
@@ -87,7 +123,6 @@ def create_show_episode_video_file(show_episode_id: int, video_file_id: int):
         db.refresh(dbm)
         return dbm
 
-
 def get_show_episode_video_file(show_episode_id: int, video_file_id: int):
     with DbSession() as db:
         return (
@@ -96,7 +131,6 @@ def get_show_episode_video_file(show_episode_id: int, video_file_id: int):
             .filter(dm.ShowEpisodeVideoFile.video_file_id == video_file_id)
             .first()
         )
-
 
 def create_show_episode_image_file(show_episode_id: int, image_file_id: int):
     with DbSession() as db:
@@ -108,7 +142,6 @@ def create_show_episode_image_file(show_episode_id: int, image_file_id: int):
         db.refresh(dbm)
         return dbm
 
-
 def get_show_episode_image_file(show_episode_id: int, image_file_id: int):
     with DbSession() as db:
         return (
@@ -117,7 +150,6 @@ def get_show_episode_image_file(show_episode_id: int, image_file_id: int):
             .filter(dm.ShowEpisodeImageFile.image_file_id == image_file_id)
             .first()
         )
-
 
 def create_show_episode_metadata_file(show_episode_id: int, metadata_file_id: int):
     with DbSession() as db:
@@ -128,7 +160,6 @@ def create_show_episode_metadata_file(show_episode_id: int, metadata_file_id: in
         db.commit()
         db.refresh(dbm)
         return dbm
-
 
 def get_show_episode_metadata_file(show_episode_id: int, metadata_file_id: int):
     with DbSession() as db:
@@ -156,41 +187,56 @@ def upsert_show_episode_tag(show_episode_id: int, tag_id: int):
 
 def get_partial_show_episode_list(ticket:dm.Ticket,season_id:int,only_watched:bool=True):
     with DbSession() as db:
-        episodes = get_show_episode_list_by_season(show_season_id=season_id)
-        season = db_season.get_show_season_by_id(season_id=season_id)
-        show = db_show.get_show_by_id(show_id=season.show.id)
-        shelf_watched = db_show.get_show_shelf_watched(cduid=cduid,shelf_id=show.shelf.id)        
+        episodes = get_show_episode_list_by_season(ticket=ticket,show_season_id=season_id)
+        if not episodes:
+            return []
+        season = db_season.get_show_season_by_id(ticket=ticket,season_id=season_id)
+        if not season:
+            return []
+        show = db_show.get_show_by_id(ticket=ticket,show_id=season.show.id)
+        if not show:
+            return []
+        shelf_watched = db_show.get_show_shelf_watched(ticket=ticket,shelf_id=show.shelf.id)        
         if shelf_watched:
             return episodes if only_watched else []
-        show_watched = db_show.get_show_watched(cduid=cduid,show_id=show.id)
+        show_watched = db_show.get_show_watched(ticket=ticket,show_id=show.id)
         if show_watched:
             return episodes if only_watched else []        
-        season_watched = db_season.get_show_season_watched(cduid=cduid,season_id=season_id)
+        season_watched = db_season.get_show_season_watched(ticket=ticket,season_id=season_id)
         if season_watched:
             return episodes if only_watched else []        
         watched_episodes = db.query(dm.Watched).filter(
-            dm.Watched.client_device_user_id == cduid,
+            dm.Watched.client_device_user_id.in_(ticket.watch_group),
             dm.Watched.shelf_id == show.shelf.id,
             dm.Watched.show_id == show.id,
             dm.Watched.show_season_id == season_id,
             dm.Watched.show_episode_id != None
-        ).all()
+        ).distinct(dm.Watched.show_episode_id).all()
         watched_ids = [xx.show_episode_id for xx in watched_episodes]
         if only_watched:
             return [xx for xx in episodes if xx.id in watched_ids]
         return [xx for xx in episodes if not xx.id in watched_ids]
 
-def set_show_episode_watched(cduid:int,episode_id:int,is_watched:bool=True):
+def set_show_episode_watched(ticket:dm.Ticket,episode_id:int,is_watched:bool=True):
     with DbSession() as db:
-        episode = get_show_episode_by_id(episode_id=episode_id)
-        season = db_season.get_show_season_by_id(season_id=episode.season.id)
-        show = db_show.get_show_by_id(show_id=season.show.id)
+        episode = get_show_episode_by_id(ticket=ticket,episode_id=episode_id)
+        if not episode:
+            return False
+        season = db_season.get_show_season_by_id(ticket=ticket,season_id=episode.season.id)
+        if not season:
+            return False
+        show = db_show.get_show_by_id(ticket=ticket,show_id=season.show.id)
+        if not show:
+            return False
         shelf_id = show.shelf.id
-        shelf_watched = db_show.get_show_shelf_watched(cduid=cduid,shelf_id=shelf_id)
-        show_watched = db_show.get_show_watched(cduid=cduid,show_id=show.id)
-        season_watched = db_season.get_show_season_watched(cduid=cduid,season_id=season.id)
-        episodes = get_show_episode_list_by_season(show_season_id=season.id)
+        shelf_watched = db_show.get_show_shelf_watched(ticket=ticket,shelf_id=shelf_id)
+        show_watched = db_show.get_show_watched(ticket=ticket,show_id=show.id)
+        season_watched = db_season.get_show_season_watched(ticket=ticket,season_id=season.id)
+        episodes = get_show_episode_list_by_season(ticket=ticket,show_season_id=season.id)
+        if not episodes:
+            return False
         db.query(dm.Watched).filter(
+            dm.Watched.client_device_user_id.in_(ticket.watch_group),
             dm.Watched.show_episode_id == episode_id
         ).delete()
         db.commit()
@@ -201,14 +247,14 @@ def set_show_episode_watched(cduid:int,episode_id:int,is_watched:bool=True):
                     dm.Watched.show_id == show.id,
                     dm.Watched.show_season_id == season.id,
                     dm.Watched.show_episode_id != None
-                ).all()
+                ).distinct(dm.Watched.show_episode_id).all()
             )
             if len(watched_episodes) == len(episodes) - 1:
-                db_season.set_show_season_watched(cduid=cduid,season_id=season.id,is_watched=True)
+                db_season.set_show_season_watched(ticket=ticket,season_id=season.id,is_watched=True)
                 return True
             else:
                 dbm = dm.Watched()
-                dbm.client_device_user_id = cduid
+                dbm.client_device_user_id = ticket.cduid
                 dbm.shelf_id = shelf_id
                 dbm.show_id = show.id
                 dbm.show_season_id = season.id
@@ -219,11 +265,11 @@ def set_show_episode_watched(cduid:int,episode_id:int,is_watched:bool=True):
                 return True
         if not is_watched:
             if shelf_watched:
-                db_show.set_show_shelf_watched(cduid=cduid,shelf_id=shelf_id,is_watched=False)
+                db_show.set_show_shelf_watched(ticket=ticket,shelf_id=shelf_id,is_watched=False)
             if show_watched:
-                db_show.set_show_watched(cduid=cduid,show_id=show.id,is_watched=False)
+                db_show.set_show_watched(ticket=ticket,show_id=show.id,is_watched=False)
             if season_watched:
-                db_season.set_show_season_watched(cduid=cduid,season_id=season.id,is_watched=False)
+                db_season.set_show_season_watched(ticket=ticket,season_id=season.id,is_watched=False)
                 episodes_watched = []
                 for other_episode in episodes:
                     if other_episode.id == episode.id:
@@ -233,22 +279,26 @@ def set_show_episode_watched(cduid:int,episode_id:int,is_watched:bool=True):
                         'show_season_id': season.id,
                         'show_id': show.id,
                         'shelf_id': shelf_id,
-                        'client_device_user_id': cduid
+                        'client_device_user_id': ticket.cduid
                     })
                 db.bulk_insert_mappings(dm.Watched,episodes_watched)
                 db.commit()
                 return False
     return is_watched
 
-def get_show_episode_watched(cduid:int,episode_id:int):
-    episode = get_show_episode_by_id(episode_id=episode_id)
-    season = db_season.get_show_season_by_id(season_id=episode.season.id)    
-    season_watched = db_season.get_show_season_watched(cduid=cduid,season_id=season.id)
+def get_show_episode_watched(ticket:dm.Ticket,episode_id:int):
+    episode = get_show_episode_by_id(ticket=ticket,episode_id=episode_id)
+    if not episode:
+        return False
+    season = db_season.get_show_season_by_id(ticket=ticket,season_id=episode.season.id)    
+    if not season:
+        return False
+    season_watched = db_season.get_show_season_watched(ticket=ticket,season_id=season.id)
     if season_watched:
         return True
     with DbSession() as db:
         watched = db.query(dm.Watched).filter(
-            dm.Watched.client_device_user_id == cduid,
+            dm.Watched.client_device_user_id == ticket.cduid,
             dm.Watched.shelf_id == season.show.shelf.id,
             dm.Watched.show_id == season.show.id,
             dm.Watched.show_season_id == season.id,
